Update bashbs.sh
updated the for loop to save the makefile correctly

+++ bashbs.sh
@@ -6,7 +6,7 @@ usage(){
 }
 pname=$1
 pdir="./${pname}"
-tdir=/var/bash.bs/templates
+tdir=/home/runner/workspace/templates
 
 
 if [ -z "$pname" ]; then
@@ -33,8 +33,13 @@ cp -R ${tdir}/$template $pdir
 cd $pdir
 for x in *; do
   new=$(sed "s,PROJECTNAME,$pname,g" <<< "$x")
-  sed "s,PROJECTNAME,$pname,g" < $x > $new
-  if [ -e "$new" ]; then
+  if [ "$x" = "$new" ]; then
+    sed "s, PROJECTNAMEA,$pname,g" < $x > temp
+    mv -f temp $x
+  else
+    sed "s,PROJECTNAME,$pname,g" < $x > $new
+    if [ -e "$new" ]; then
       rm -f $x
+    fi
   fi
 done
